test float

+++ AngularData.cpp
@@ -29,17 +29,17 @@ AngularData::AngularData(int maxk) {
 				for (int m2 = -l2; m2 <= l2; m2++, lm2++) {
 					for (int i = 0; i<3; i++) {
 						for (int j = 0; j<3; j++) {
-							REAL *r1 = (i==0)?q.x:(i==1)?q.y:q.z;
-							REAL *r2 = (j==0)?q.x:(j==1)?q.y:q.z;
-							REAL sum = 0;
+							double *r1 = (i==0)?q.x:(i==1)?q.y:q.z;
+							double *r2 = (j==0)?q.x:(j==1)?q.y:q.z;
+							double sum = 0;
 							for (int p = 0; p < q.order; p++) {
-								REAL t1 = s.value(l1, m1, q.x[p], q.y[p], q.z[p]);
-								REAL t2 = s.value(l2, m2, q.x[p], q.y[p], q.z[p]);
+								double t1 = s.value(l1, m1, q.x[p], q.y[p], q.z[p]);
+								double t2 = s.value(l2, m2, q.x[p], q.y[p], q.z[p]);
 								sum += q.w[p] * t1 * t2 * r1[p] * r2[p];
 							}
 							if ((fabs(sum) > 1e-10) || j == 2) {
 							//	prinf("l1 = % d l2 = % d m1 = % d m2 = % d i = % d j = % d s = %e", l1, l2, m1, m2, i, j, sum);
-								omega[i*aslm*aslm + aslm * lm1 + lm2] = sum;
+								omega[i*aslm*aslm + aslm * lm1 + lm2] = (REAL)sum;
 								omega_pos[i*aslm*aslm + aslm * lm1 + lm2] = j;
 								break;
 							}
